Clean up example applications
This change cleans the example applications provided to show
some simple routes with camel-milo.
